Move generation_keeper.py
