+++ fedot/core/optimisers/generation_keeper.py
