+++ fedot/core/optimisers/gp_comp/parameters/population_size.py
@@ -2,7 +2,7 @@ import math
 
 from .parameter import AdaptiveParameter
 from fedot.core.utilities.data_structures import BidirectionalIterator
-from fedot.core.optimisers.gp_comp.generation_keeper import ImprovementWatcher
+from fedot.core.optimisers.generation_keeper import ImprovementWatcher
 
 PopulationSize = AdaptiveParameter[int]
